requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
langchain_chroma
langchain_google_genai